Accept access token in construction for hydration. Add invalid callback
This allows the consuming application to only store the access token and it will call out to the service to get all of the facts (such as the refresh token).

Not only will it allow a hard reload to refresh the local cache, but increase safety by not expecting the consuming app to store the refresh token somewhere.

The invalid callback allows the app to also clear its storage of anything in the event the constructor's hydration call fails due to expiration or a regular call failure due to revocation