Abstract Promise in constructor, block HTTP calls until its finished
Previously v1.1.0 leaked abstraction by returning a Promise from the constructor. This is not only buggy/not allowed in Node world, but is also a bad design.

Instead, the constructor still fires a promise to load the context remotely (when an access token is provided) and any subsequent HTTP calls will wait until its resolved before continuing.