Remove using only access token to get refresh token and context
This bootstrapping isn't realistic as it relies on a violation of OAuth2 spec: the info endpoint does not reveal the refresh token.

Constructor will still verify the authenticity of the token but wont clobber its own context with the response. If it fails, the client will pivot to using the refresh token first to recover access.