Store and expect simple-oauth2 objects in auth object
We no longer rely on the info endpoint to update the auth object, just simply run the revoked callback if not valid.

Because of our custom headers and use of fetch, we wont use simple-oauth2's HTTP calls, but take advantage of its access token object and helpers.

This makes the client handle refresh tokens appropriately :)